Refactor test display names and rearrange test methods
Refactored display names of test methods in FizzBuzzFilterExerciseTest.java for better readability and consistency - the task number was added in front of each name. Removed an imported but unused class, IntStream, to improve code cleanliness.

Additionally, a test method concerning the length of the sequence was rearranged and placed right above another method for similar purpose, thus improving the coherence of code. Added detailed doc for the flipFlop method to explain its functionality for easier future reference.