Modify variable declaration and format code in README.md
Altered explicit type declaration to "var", making the code cleaner and more modern as per the introduction of JEP 286 in JDK 10. Also, formatted the pieces of code and examples to improve readability, remove side distractions, and keep the focus solely on the logic. This modification does not change any operation or output of the code and tools, purely cosmetic.